requirements.txt
fastapi[all]
sqlalchemy
psycopg2
bcrypt==4.0.1
passlib
python-jose[cryptography]
stripe